Moved import notify to its own service in order to separate specific from general logic
