+++ app/common/notifications/notify.service.js
@@ -14,8 +14,7 @@ function Notify(_, $q, $rootScope, $translate, SliderService, ModalService) {
         confirm: confirm,
         confirmModal: confirmModal,
         confirmDelete: confirmDelete,
-        limit: limit,
-        importComplete: importComplete
+        limit: limit
     };
 
     function notify(message, translateValues) {
@@ -186,18 +185,6 @@ function Notify(_, $q, $rootScope, $translate, SliderService, ModalService) {
         function showSlider(message) {
             scope.message = message;
             SliderService.openUrl('templates/common/notifications/limit.html', 'warning', 'error', scope, true, false);
-        }
-    }
-
-    function importComplete(translateValues) {
-        var scope = getScope();
-
-        scope = _.extend(scope, translateValues);
-
-        showSlider();
-
-        function showSlider() {
-            SliderService.openUrl('templates/common/notifications/import-complete.html', 'thumb-up', 'confirmation', scope, false, false);
         }
     }
