+++ app/settings/data-import/import.notify.service.js
@@ -1,0 +1,30 @@
+module.exports = ImportNotify;
+
+var scope;
+
+ImportNotify.$inject = ['_', '$q', '$rootScope', '$translate', 'SliderService', 'ModalService'];
+function ImportNotify(_, $q, $rootScope, $translate, SliderService, ModalService) {
+    return {
+        importComplete: importComplete
+    };
+
+    function importComplete(values) {
+        var scope = getScope();
+
+        scope = _.extend(scope, values);
+
+        showSlider();
+
+        function showSlider() {
+            SliderService.openUrl('templates/common/notifications/import-complete.html', 'thumb-up', 'confirmation', scope, false, false);
+        }
+    }
+
+    function getScope() {
+        if (scope) {
+            scope.$destroy();
+        }
+        scope = $rootScope.$new();
+        return scope;
+    }
+}
