+++ app/settings/settings.module.js
@@ -3,6 +3,7 @@ angular.module('ushahidi.settings', [])
 .directive('importerCsv', require('./data-import/data-import.directive.js'))
 .directive('mapperCsv', require('./data-import/data-mapper.directive.js'))
 .directive('configureCsv', require('./data-import/data-configure.directive.js'))
+.service('ImportNotify', require('./data-import/import.notify.service.js'))
 
 .directive('surveyEditor', require('./surveys/survey-editor.directive.js'))
 .directive('surveyTaskCreate', require('./surveys/task-create.directive.js'))
